+++ branch/src/main/scala/dev/alteration/branch/hollywood/rag/Document.scala
@@ -1,0 +1,3 @@
+package dev.alteration.branch.hollywood.rag
+
+case class Document(id: String, content: String, embedding: List[Double])
