+++ branch/src/main/scala/dev/alteration/branch/hollywood/rag/ScoredDocument.scala
@@ -1,0 +1,4 @@
+package dev.alteration.branch.hollywood.rag
+
+case class ScoredDocument(document: Document, score: Double)
+
